refactor: modularize SlotTable component into reusable parts
