+++ client/src/components/index.ts
@@ -1,4 +1,4 @@
-export { default as SlotTable } from './SlotTable';
+export { default as SlotTable } from './slots/SlotTable';
 import AppLogo from './AppLogo';
 
 export { AppLogo };
