+++ client/src/components/slots/index.ts
@@ -1,0 +1,5 @@
+import SlotTable from './SlotTable';
+import TimeSlotRow from './TimeSlotRow';
+import TableHeader from './TableHeader';
+
+export { SlotTable, TimeSlotRow, TableHeader };
