+++ client/src/components/slots/styled.ts
@@ -1,0 +1,19 @@
+import { styled, TableCell, tableCellClasses, TableRow } from '@mui/material';
+
+export const StyledTableCell = styled(TableCell)(({ theme }) => ({
+  textAlign: 'center',
+  border: '1px solid rgba(224, 224, 224, 1)',
+  [`&.${tableCellClasses.head}`]: {
+    backgroundColor: theme.palette.primary.main,
+    color: theme.palette.common.white,
+  },
+  [`&.${tableCellClasses.body}`]: {
+    fontSize: 14,
+  },
+}));
+
+export const StyledTableRow = styled(TableRow)(({ theme }) => ({
+  '&:nth-of-type(odd)': {
+    backgroundColor: theme.palette.action.hover,
+  },
+}));
